Polishing: small renaming and Typedoc
