+++ src/spi/deploy/Deployment.ts
@@ -1,7 +1,13 @@
 import { RemoteRepoRef } from "@atomist/automation-client/operations/common/RepoId";
 
+/**
+ * Information about a deployment
+ */
 export interface Deployment {
 
+    /**
+     * The root endpoint, if the deployment has an endpoint
+     */
     readonly endpoint?: string;
 }
