+++ src/spi/deploy/SourceDeployer.ts
@@ -1,13 +1,14 @@
 import { ProjectOperationCredentials } from "@atomist/automation-client/operations/common/ProjectOperationCredentials";
 import { RemoteRepoRef } from "@atomist/automation-client/operations/common/RepoId";
-import { AddressChannels } from "../../common/slack/addressChannels";
 import { ManagedDeploymentTargetInfo } from "../../handlers/events/delivery/deploy/local/appManagement";
 import { LogInterpreter } from "../log/InterpretedLog";
 import { ProgressLog } from "../log/ProgressLog";
 import { Deployment } from "./Deployment";
 
 /**
- * Implemented by classes that can deploy from source
+ * Implemented by classes that can deploy from source.
+ * Such deployers do not need an artifact to have been built by
+ * execution of a previous Build goal.
  */
 export interface SourceDeployer {
 
@@ -19,7 +20,7 @@ export interface SourceDeployer {
     deployFromSource(id: RemoteRepoRef,
                      ti: ManagedDeploymentTargetInfo,
                      log: ProgressLog,
-                     creds: ProjectOperationCredentials,
+                     credentials: ProjectOperationCredentials,
                      atomistTeam: string): Promise<Deployment>;
 
     logInterpreter?: LogInterpreter;
